requirements.txt
discord.py
flask
requests
firebase-admin
python-dotenv